Reduce default TTL to 60s
This was modified in live config #49. In response to Number 10 issue.
Additional detail from Brad:

The timeout Fastly sets for 302 redirects is unclear, and it appears to be
larger than Akamai's. This caused the 302 for virus-checking placeholder to
be cached downstream which broke the featured image on their homepage.

To aid this in the future I've added an explicit timeout to the virus image
placeholder controller action in the app itself.