Restrict Fastly HTTP purges by IP
Fastly's default behaviour for purging is to allow anyone (yes, anyone!) to
purge objects stored on their cache nodes. For obvious reasons, this isn't too
good - we can't be granting anyone the ability to remove objects from their
nodes as it can lead to utter chaos.

To that end, this commit defines an ACL of approved IPs, then allows the use of
the FASTLYPURGE verb by those IPs alone. Any request to purge from an IP outside
of this ACL returns a HTTP 403.

The list of IPs in the ACL was chosen specifically because there are
requirements to getting access to machines that use them. Given that the ability
to purge objects can cause undesirable side-effects, we restrict this ability to
IPs which require production access. The Aviation House IPs were considered to
add to this list, however as non-technical users have access, this was not
deemed a good idea. Similarly, the guest WiFi APs use the same IPs as the staff
APs - therefore allowing, potentially, guests to purge objects.

Setting the HTTP verb to GET allows us to return the 403 error without any
interference from other VCL. If this is not included, the 403 error is not
returned when it should be.

The order of the codeblock in the VCL was also considered. It is important that
the FASTLYPURGE IP-based restrictions codeblock is above the Force SSL codeblock
in order to ensure that they are not bypassed if the Force SSL codeblock needs
to return an error.

@dancarley and I have tested this and it appears to be working well.

\o/