Make Redirector the special case, not Bouncer
We had some tech debt here in that the if-statement it was generating was far
far bigger than it needed to be - we now have thousands of hostnames going to
Bouncer, and only a few going to Redirector.

The Varnish configuration that Fastly were ending up with was invalid because
it was truncated part-way through the if-statement after we recently added
numerous hosts to Bouncer.