Set origin probe frequency to 1 second
Set the frequency of the origin backend health check probe so that it
runs every second.

Note that this is an interval and not a rate. The rate at which the
probe will check the backend is determined as[1]:

interval + timeout = rate

This complements the change in commit
ca5868bffabf70ad8830a83fc1bc11f89e73e728. By checking the backend health
more frequently, Varnish will be faster to detect if origin is sick and
will start serving from stale which is preferable to serving from the
mirrors.

References [#57975786].

[1]: https://www.varnish-cache.org/trac/wiki/BackendPolling#a.interval