Serve from stale if obj in grace and origin 5XXs
Serve from stale if the requested object is in the defined grace period
(currently set to 24 hours) if origin returns a 5XX error. This
implements [#57975786].

Previously, if we received a 5XX error we would have served from the
first mirror. Serving from stale is preferable because:

- Stale objects will be more current than the content served by a mirror
- The mirrors are only updated every 24 hours, so may be missing new
content
- The mirrors do not contain any dynamic content such as search results

This should reduce the number of friendly 503 errors we serve if search
is unavailable and the search results are cached.

This also uses Varnish's saintmode to prevent us from requesting the
same URL against origin for the next 60 seconds. Any requests received
during that time will be served either from stale if cached, or from the
mirrors. This implements [#58345516].

I've also changed the units for req.grace from seconds to hours to make
it more readable.