Serve from mirror if origin is down, probe is OK
Allows Varnish to restart the request if origin is not responding but
the backend probe for origin has not yet detected that origin is sick.

This works using a slightly hackish workaround that sets a custom
header, X-VCL-Error, in vcl_error() which is then acted upon in
vcl_recv().

We serve from the mirrors in this case because, as evidenced in our
testing[1], Varnish won't serve from stale if it thinks that the backend
is healthy.

This implements [#57975674] because it will prevent us from serving
Varnish's built-in 'Guru Meditation' error, unless all our backends are
unavailable and there is no stale object we can serve.

Note that vcl_error() will attempt to restart 3 times so that the logic
in vcl_recv() can try stale then each of the backends in turn until the
request is served.

[1]: See [this](serve-from-stale-004.vtc) test introduced in
86030bbdfcc91a8fbd95ab51608b650104dd1ced.