Correct failover logic to serve from stale
Correct the failover logic to mandate more than one restart before
attempting to use the first mirror backend.

This is essential to allow Varnish to serve from stale even if
vcl_error() has been triggered or if origin is unhealthy.

Before this fix, Varnish would immediately failover to the first mirror
as soon as the origin backend was determined to be sick by its probe,
therefore Varnish would not serve from stale.

This explains the behaviour seen after the probe interval was reduced to
1 second in 8498143ad38f0290aae8a40c4af49768b6a72f6c.

Implements [#57975786].