Simplify failover logic
Simply the failover logic by relying solely on restarts.

Using Varnish restarts makes the failover logic much cleaner and allows
us to retain the same behaviour in case a backend is down or a HTTP 5XX
response is received.

This change may cause one extra restart than the previous logic - if
origin is unhealthy, Varnish will hit vcl_error() and restart rather
than switching to stale or the mirrors immediately as it did previously.
I think this is an acceptable compromise for cleaner code.

Note that the backend probes still serve a purposes as without them,
Varnish will not serve from stale; req.grace does not work on healthy
backends.