Set saintmode to 5s
'saintmode' prevents Varnish from resending a request to the same
backend within the specified period of time.

We set saintmode when vcl_fetch() receives a HTTP 5XX error. These
errors may only be temporary blips, due to an application timeout during
busy periods for example.

By setting saintmode to 5 seconds, we increase the likelihood that
end-users will retry the request against origin. An end-user is unlikely
to wait 60 seconds before retrying a page.

It's important that we retry against origin rather than relying on the
mirrors as dynamic content, using GET query string parameters for
example, will not be stored by the mirrors.