Remove PURGE req.request hack
Remove the hack that Fastly advised us to do in 4ce71d6 which ensured that
the correct response (status and body) was returned to the client.

A change in behaviour at their end, between the time we did that and when we
were able to test it again with cdn-acceptance-tests, meant that it was
transforming the request into a GET which went to origin and returned a
normal body. As shown by:

    === RUN TestMiscRestrictPurgeRequests
    --- FAIL: TestMiscRestrictPurgeRequests (0.12 seconds)
            cdn_misc_test.go:58: Request should not have made it to origin
            cdn_misc_test.go:75: Request 2 received incorrect status code. Expected 403, got 200

With this change the test now passes:

    === RUN TestMiscRestrictPurgeRequests
    --- PASS: TestMiscRestrictPurgeRequests (0.11 seconds)