Remove XFF handling from vcl_recv
This code was copied from a Fastly rendered VCL when the service was still
configured from the GUI. It relates to an option visible in their GUI diff
option as:

    request_settings:
    - name: Default
      …
      xff: append

It has been causing the client IP to be appended to the XFF header twice.
This behaviour might have changed after we copied it, or maybe it was always
been this way and we never looked:

    === RUN TestHeaderXFFCreateAndAppend
    --- FAIL: TestHeaderXFFCreateAndAppend (0.00 seconds)
            cdn_acceptance_test.go:86: Expected origin to receive "X-Forwarded-For" header with single IP. Got "80.240.128.97, 80.240.128.97"

Looking at the Production Assets service which is still configured by the
GUI and has this option enabled shows that it no longer renders this code in
the VCL. Removing it fixes the test, so I think that's the right thing to
do:

    === RUN TestReqHeaderXFFCreateAndAppend
    --- PASS: TestReqHeaderXFFCreateAndAppend (0.22 seconds)

We can also get rid of the spurious empty condition and the comment about
it. Fastly said this wasn't a typo:

- https://fastly.zendesk.com/requests/2251

But it looks suspiciously like a copy/paste error from `default.vcl`, hey ho:

- https://github.com/varnish/Varnish-Cache/blob/2.1/bin/varnishd/default.vcl#L43-50