Don't cache Authorization/Cookie reqs
Don't cache responses to requests which have Authorization or Cookie headers
because they are likely to contain data that is personalised or private to
the original client.

This won't affect our hit/miss ratio for www.gov.uk because no frontend apps
in production use BasicAuth and the only frontend app that uses cookies is
Licensing. We explictly strip inbound and outbound cookies at origin:

- https://github.gds/gds/puppet/blob/1edeb5b396bb09c3922350a3d12283cfce2a8f8b/modules/varnish/templates/default.vcl.erb#L66-72
- https://github.gds/gds/puppet/blob/1edeb5b396bb09c3922350a3d12283cfce2a8f8b/modules/varnish/templates/default.vcl.erb#L103-106

This makes these two cdn-acceptance-tests now pass:

    === RUN TestNoCacheHeaderAuthorization
    --- PASS: TestNoCacheHeaderAuthorization (0.04 seconds)
    === RUN TestNoCacheHeaderCookie
    --- PASS: TestNoCacheHeaderCookie (0.03 seconds)

The code has been copied from Varnish's default vcl_recv behaviour:

- https://github.com/varnish/Varnish-Cache/blob/2.1/bin/varnishd/default.vcl#L65-68

We tried removing the explicit `return(lookup)` so that Fastly would fall
through to this, as Varnish does by default, however this reverted our fix
for [#73192066] by appending to XFF twice because of this:

- https://github.com/varnish/Varnish-Cache/blob/2.1/bin/varnishd/default.vcl#L43-50

That isn't such a bad thing though. Because we already have a variation of
the other condition in that file and we don't know what's in Fastly's
`default.vcl` so the less we depend on it the better.