Revert req.http.Cookie return(pass)
This caused a huge increase in traffic to origin when deployed to prod.

In hindsight, it occured to us that, although we strip most `Set-Cookie`
headers from responses at origin, cookies are still set by Javascript and
their presence causes the browser to send them with every request.

So revert this change while we figure out the correct behaviour.