Rate-Limit-Token header for prod/stag
Add a Rate-Limit-Token header to health check probes for the production and
staging service. This will allow us to increase the probe interval without
it hitting rate limiting on our `cache-N.router` nodes.

The condition means that this header won't be present for the acceptance
test service where we don't need to do this. We also don't need to do this
for the mirrors, because they don't do any rate limiting.