Reduce probe intervals to 10s
So that they detect failure conditions sooner. This is possible now that we
pass Rate-Limit-Token to bypass rate limiting at origin.

We may want to experiment with reducing this interval further in the future.
But this is a good start and puts us in parity with what we've been using
for the acceptance test service.