Revert "[#73121560] Don't cache Authorization/Cookie reqs"
This reverts commit cf6d76725a120832c78f756f3ec62866ebeb0998.

Despite the RFC stating that responses to requests with an Authorization
header should also opt-in to caching it's not the default behaviour in
Fastly or CloudFlare. It has also raised our number of "pass" requests
because Smokey unnecessarily passes authentication in production - not a bad
thing in itself because it already cache busts with query params, but it's
not a good side effect. I'll adjust the cdn-acceptance-tests to reflect
this.

http://tools.ietf.org/html/rfc7234#section-3.2

Conflicts:
	vcl_templates/www.vcl.erb