Override default.vcl for Set-Cookie
Explicit `return(pass)` for backend responses that have a Set-Cookie header
was removed in 64bc44e. However cdn-acceptance-tests identified that this
had no effect, I believe because we don't have an explicit `return()` at the
end of `vcl_fetch{}` which causes it to fall into Varnish's default.vcl that
performs the action as the code that we deleted:

https://github.com/varnish/Varnish-Cache/blob/2.1/bin/varnishd/default.vcl#L111-L113

Prevent this from happening (but don't prevent the fall throught to
default.vcl) by explicitly calling `return(deliver)` instead. This needs to
go at the end of `vcl_fetch{}` so that the other logic about applying
default TTLs is still executed for such responses.

Confirmed working by the tests when deployed to the acceptance test service.