Use sick backend to force stale
As used in the CDN acceptance test mocks:

- alphagov/cdn-acceptance-tests@c057b38c6c59b317cd52bb67115b28a774317195

This was failing when origin is down and the health check hasn't noticed
yet, Varnish would:

- reattempt the request against origin on the first restart
- attempt the request against the backup server on the second restart
- invalidate the stale object with the response from the backup server

We can workaround this by using the "evil hack" described here:

- https://www.varnish-software.com/static/book/Saving_a_request.html#example-evil-backend-hack

It creates a new backend which will always be "sick"; never receives a real
request. It uses a port which shouldn't be bound on the guest, a request
that isn't valid HTTP, an interval that means it will rarely be polled, and
an initial value which means that it's immediately sick when Varnish first
starts.

This causes Varnish to use the stale/grace object (if it exists) on the
first restart and never invalidate it with a response from a backup server.

Confirmed working against acceptance test service with updated tests:
```
dcarley@ci-slave-4:~/cdn-acceptance-tests$ go test -edgeHost fastly.cdn-acceptance-test.alphagov.co.uk -skipVerifyTLS -vendor fastly -v -run Stale
2014/09/24 09:32:23 Confirming that CDN is healthy
2014/09/24 09:32:23 Started server on port 8082
2014/09/24 09:32:23 Checking health of backup2...
2014/09/24 09:32:38 backup2 is up!
2014/09/24 09:32:38 Started server on port 8081
2014/09/24 09:32:38 Checking health of backup1...
2014/09/24 09:32:49 backup1 is up!
2014/09/24 09:32:49 Started server on port 8080
2014/09/24 09:32:49 Checking health of origin...
2014/09/24 09:33:00 origin is up!
=== RUN TestServeStaleOriginDown