Return a pass when max-age=0
Fastly don't respect the `Cache-Control: max-age=0` header when we set it in
our CDN acceptance tests. We set this to ensure that we can test that
requests are not cached when the header is set. See
https://github.com/alphagov/cdn-acceptance-tests/blob/1808d06890d58cbf6d1339cbce3b2dc6ed8c1c57/cdn_nocache_test.go#L83-L93

When multiple requests with that header are set, then whilst the first
response comes back as expected - with the correct response - future
requests also come back with the same response as the first request, when
they should come back with the responses from the second and third requests
respectively.

When asked, in http://fastly.zendesk.com/requests/8253, Fastly advised us to
return a `pass` in VCL when this header exists.

This does that.