Fastly assets: remove dupe default TTL condition
This is surplus to the condition at the end of `vcl_fetch` which applies a
default TTL if the response doesn't have an `Expires` header or
`Cache-Control` header with a `max-age` value.

It was modified slightly in [#81663690] to change the condition from `if
(beresp.ttl <= 60s)` to what I'm removing now, so that it respected existing
`Cache-Control: max-age=N` headers. Whereas removing it then should have had
the same effect.