Fastly assets: allow PURGE from whitelist IPs
I've since discovered that we have purged items from the assets service in
the past. There's a fabric task that does this, although it doesn't have
much effect at the moment since we moved this service to CloudFlare:

https://github.com/alphagov/fabric-scripts/blob/a9bcded6860f59a99332141c147e3c190db7de59/cdn.py#L9

So add egress IPs of our trusted environments so that this will work if
required in the future. I'm not adding Fastly's own IPs, because they are
difficult to maintain and we shouldn't be purging from the UI.

Tested this by deploying to the acceptance test service and issuing `PURGE`
requests from the office (denied) and production (successful).