Fastly service and VCL template for TLD redirect
This is a service that we're moving from CloudFlare to Fastly. It handles
traffic for `http://gov.uk` and `https://gov.uk`, always serving a 301
redirect to `https://www.gov.uk`. The path and query params of the original
request are preserved in the redirect but fragments are not (per the spec).

I've implemented this by using as little VCL as possible. We must have a
backend because Fastly (and maybe Varnish) complains if we don't, so we
create a dummy one. All requests are immediately passed to `vcl_error` which
always serves a 301 response.

Matt and I talked about adding a test for this to cdn-acceptance-tests but
decided against it because it's too specific to our use case and too
different from the existing tests, in that:

- we'd need to test both HTTP and HTTPS
- the destination URL is always hardcoded to the `www.gov.uk` domain
- we can't do the initial healthcheck because we have no backend

So we're just going to test this by using `curl` as we did for CloudFlare
and the Puppet managed project prior to that. Which looks like this, prior
to us adding the correct domain and changing DNS to the AnyCast addresses:

    ➜  fastly git:(78447286-fastly_tld_redirect) curlh --resolve govuk-redirector.com:80:23.235.33.144 http://govuk-redirector.com/foo\?bar\#baz
    HTTP/1.1 301 Moved Permanently
    Server: Varnish
    Retry-After: 0
    Location: https://www.gov.uk/foo?bar
    Content-Length: 0
    Accept-Ranges: bytes
    Date: Thu, 05 Feb 2015 09:30:37 GMT
    X-Varnish: 1448849271
    Via: 1.1 varnish
    Connection: close

    ➜  fastly git:(78447286-fastly_tld_redirect) curlh -k --resolve govuk-redirector.com:443:23.235.33.144 https://govuk-redirector.com/foo\?bar\#baz
    HTTP/1.1 301 Moved Permanently
    Date: Thu, 05 Feb 2015 09:30:39 GMT
    Server: Varnish
    Retry-After: 0
    Location: https://www.gov.uk/foo?bar
    Content-Length: 0
    Accept-Ranges: bytes
    X-Varnish: 1642191129
    Via: 1.1 varnish

I'm also omitting a staging environment for the moment purely on the basis
that I don't know what hostname we'd give it. We can add one later if we
need to test changes.