Assets: Support failover to mirrors
Configure our Fastly VCL for the assets service to support failover to
our mirrors at Skyscape and Carrenza if the assets origin is
unavailable.

This configuration is purposefully copied from the VCL for our main
Fastly service (`vcl_templates/www.vcl.erb`) so as to maintain
consistency. There are a number of small changes as a result which aim
to make the services more consistent.

This commit includes logic for serving from stale. While we're not sure
that the serve-from-stale functionality works consistently (see
 #80276874), I think it's worth including here to maintain consistency
with our main service. We also have no reason to believe that including
it has an any adverse effects.

I have included Fastly's IPs in the `purge_ip_whitelist` ACL, because
although purging from the assets service is not something we have done
before (as mentioned in 29ff0476), I think it's worth doing to reduce
the inconsistencies between this and the VCL in
`vcl_templates/www.vcl.erb`.

Note that we now consider HTTP status codes from 500 to 599 as errors
(before it was 500 to 503); again this is for consistency.

Also, we still serve a HTML page if all backends are down because it's
still better than serving Fastly's 'Guru Mediation' error (which is also
HTML).