Allow PURGE requests to go pass through to origin
In beb7a9d, we added the VCL configuration for our assets service, which
was copied verbatim from the rendered VCL visible in the Fastly UI.

When I compared the VCL for assets to the VCL in `www.vcl.erb`, I found
that the assets VCL (as configured by Fastly) allows HTTP requests using
the `PURGE` verb (note NOT `FASTLYPURGE`) to be passed through to the
origin.

To fix this inconsistency, change `www.vcl.erb` to match the logic in
the assets VCL (which was originally determined by Fastly).

This should allow `PURGE` requests to go to origin. We don't use that
functionality currently (`PURGE` requests are [whitelisted by IP][] on our
Varnish instances on the `cache-?.router` machines), but it makes our
service configuration consistent. I don't know that passing through
`FASTLYPURGE` requests, as was the case previously, had an
adverse effect on purging content - it seemingly worked for us as it
was.

[whitelisted by IP]: https://github.gds/gds/puppet/blob/7ac5c0d8427a2903bb66b89f51fdedf11ba2f78c/modules/varnish/templates/default.vcl.erb#L76-L77