Catch 801 status codes in vcl_error
Fastly will raise an 801 HTTP status code to force SSL which should
never escape the VCL.

I believe that since we are now returning a synthetic response inside
VCL error we have to implement this logic. I think that this logic
was run by Fastly before we added a synthetic response.