Add Performance Platform VCL
Adds VCL for the Performance Platform with a backend set to a different
hostname per the values in the deploy_vcl script in the root of this repo.

Specific new logic in this works out the HTTP request method being used, and
routes the request to a specific app based on the request method. POST, PUT
and DELETE are sent to Write Backdrop. GET, HEAD and OPTIONS are sent to
Read Backdrop. If another request type is used, we return a HTTP 405:
"Method not allowed".