Update Fastly VCL for new Carrenza environments
Update the hostnames to use the new `publishing.service.gov.uk` DNS
suffix, and simplify the VCL where possible as we no longer need to
change the HTTP Host header for Staging.

The old Staging environment was intentionally configured to think it was
Production. The new Staging environment considers itself to be Staging,
hence why we no longer need to adjust the HTTP Host header for Staging.