Remove styled synthetic for non-HTML backends
The performanceplatform backend serves JSON to clients. There's no
need to serve a styled error here because nothing will ever see it.

The assets backend serves CSS, JavaScript and images that support
the HTML on GOV.UK. There's no need to serve a styled HTML synthetic
error because it will never be displayed in a browser.