Return a HTTP 405 for all OPTIONS requests
Currently, we respond to all OPTIONS requests with a 404. There are two
problems with doing this:

  1. An OPTIONS request is legitimate, and so returning a 404 is not the
     best way to handle them.

  2. Our CDN logs become filled with extraneous entries for OPTIONS when we
     look only for 404s. This makes it harder to find legitimately occurring
     404s, because there's quite a lot of noise.

HTTP 405 is designated for use when a request is received that is
legitimate, but the server does not wish to respond as the particular method
used against the particular URI is disallowed.

Additionally, send a HTTP Allow header in response to an OPTIONS request
which details that we accept DELETE, GET, HEAD, POST and PUT. It is then up
to applications to work out which HTTP verbs they will be serving, and which
they will not be.

Doing this will reduce noise in the logs and allow us to conform to HTTP
RFCs more.