Set `Fastly-Backend-Name` HTTP header
Fastly provides a value called `beresp.backend.name` which tracks
the backend which the response originated from:

https://docs.fastly.com/guides/performance-tuning/tracking-your-origins-name-ip-and-port

However, the value of `beresp.backend.name` that Fastly provides includes
the service ID, which Fastly considers sensitive:

> 12:40:15 <alexmuller>
> Hey, I have a question about service IDs. Should we consider them secret?
> As far as I know they can't be used to access any information without authentication.
>
> 12:42:53 <+orangeacme>
> alexmuller: we tend to consider them sensitive. best not to post in public IRC/GitHub repo

We could do a `regsub()` to remove the service ID, eg:

```
set beresp.http.Fastly-Backend-Name = regsub(beresp.backend.name, "^<%= config['service_id'] %>--", "");
```

But this is brittle.

So instead we set the header to a fixed string whenever we change the value of `req.backend`.

By setting this value to an HTTP header we can access it in 2 places:

1) The HTTP response that's served via Fastly to end users, which is helpful for troubleshooting
2) The remote syslog that we have set up with this format: `%h %l %u %t %r %>s resp.http.Fastly-Backend-Name`