Allow TLS certificate validation in Fastly config
Fastly now supports TLS certificate validation:

https://docs.fastly.com/guides/securing-communications/connecting-to-origins-over-tls#specifying-acceptable-tls-protocol-versions

Validation can include:

a. verifying the start and expiration dates of the certificate
b. verifying the Common Name and Subject Alternative Name(s)
c. specifying a minimum TLS version (e.g. 1.2)
d. setting acceptable cipher suites
e. setting the CA (e.g. self-signed if desired)
f. setting a client certificate (i.e. to authenticate requests as
   originating from Fastly)

The change in this commit supports points a-d. I've intentionally set
strict (or 'more secure') defaults, except for the cipher list which
would be difficult to maintain as a default value - it's preferable to
set this per-service when required (which we intend to for GOV.UK).

Point (e) is desirable to implement, but could be somewhat brittle and I
think the risk of an outage due to a mismatch when renewing certificates
outweighs the benefit, assuming Fastly's certificate authority bundle
is trustworthy. Either way, we can review that in future.

We can't implement point (f) because it would make it difficult for us
to test our origin from the GDS office.

The values set here in VCL correspond to the UI settings described in
the Fastly documentation linked to above. The values were described to
us in support ticket #19868:

https://fastly.zendesk.com/hc/en-us/requests/19868

> Here is the VCL for some basic TLS settings (along with sample
> values):

    .ssl = true;
    .ssl_check_cert = always;
    .ssl_cert_hostname = "example.cert.hostname.com";
    .ssl_sni_hostname = "example.sni.goes.here.com";

> If you're looking to set the min/max TLS versions, those VCL commands
> are (along with a sample value):

    .max_tls_version = "1.2";
    .min_tls_version = "1.0";
    .ssl_ciphers = "ECDHE-RSA-AES128-GCM-SHA256:!RC4";

> Finally for self-signed certificates:

    .ssl_ca_cert = {"-----BEGIN CERTIFICATE-----
    <SNIPPED PEM CONTENT>
    -----END CERTIFICATE-----"};

`ssl_cert_hostname` and `ssl_sni_hostname` supersede `ssl_hostname`,
which we used previously. Fastly have explained this thus:

> This is a historic parameter we're working on deprecating. It predates
> both ssl_cert_hostname and ssl_sni_hostname and essentially acted as a
> combination of the two - e.g. setting ssl_hostname to foo.bar would
> send foo.bar as the SNI hostname in the TLS client hello and would
> also enforce that the certificate had foo.bar as a subject CN or SAN.
> We ran into circumstances where customers wanted to change the default
> certificate hostname but could not support SNI, so we decided to split
> the parameter moving forward.