Remove `dummy` from Bouncer VCL
This has been included in this VCL ever since the first time this
file was committed. I can't find any docs on what it does but from
testing it seems to make the probe ignore the response code from
the backend, which is not behaviour that we want.

I asked about it in the #fastly IRC channel on Freenode and got
this response:

> 10:09:31 <+justin|fastly> hey alexmuller. .dummy means the healthcheck isn't currently in use