Fastly VCL: Make timeouts and interval consistent
Our timeouts when connecting to the origin are different in our `www`,
`assets` and `performanceplatform*` services.

Make them consistent, using these values:

  first_byte_timeout = 15s
  between_bytes_timeout = 10s

These seem more appropriate than the previous values of 120s, used for
for both timeouts, in the `www` service:

- 15 seconds is long enough to make a user wait (referring to the first
  byte timeout); in that case it's probably best to failover to the
  mirror

- Nginx, which we use at origin, already buffers proxied requests from
  upstreams, so it's unlikely that more than 10 seconds would pass
  between bytes for reasons other than networking:
  http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_request_buffering

Also set the probe interval to 10 seconds, which seems more sensible
than the previous interval of 60 seconds used by the `assets` service.
This should allow us to detect when origin is unhealthy more quickly.
We did the same for the `www` service in 574e48b0.