Pre-empt request changes in vcl_recv before purge
The previous logic for accepting a purge only from a predefined whitelist of
IP addresses was potentially flawed in that vcl_recv might further modify an
acceptable FASTLYPURGE request before it is performed at the edge, yet the
current logic would return a lookup immediately after this logic had
applied. This meant that if the request was ever changed in vcl_recv, it
would be returned instantly rather than the remainder of logic in the VCL
being performed.

Fixes this in order to futureproof vcl_recv to some extent.