Make A/B cookie checks consistent
Change the syntax of the A/B cookie check in vcl_recv the same as in
vcl_deliver. When we tested this on Staging, they appeared to do the
same thing, but it's confusing for them to be inconsistent and it's
possible that the old syntax didn't handle a Cookie header with multiple
cookies correctly.