Invert the A/B split logic to improve readability
This change flips the A/B fraction calculation so it calculates the
fraction of users in the B group, to be consistent with how we talk
about A/B tests.

We usually talk about A/B tests in terms of the fraction of users who
see the B version, which is the newer version. So "0%" would mean the
experiment hasn't started yet, and "100%" would mean that everyone is
being shown the new version.