Make the raw domain HTTPS upgrade HSTS compatible and set 10 second max-age
This is the first 2 steps in https://github.com/alphagov/govuk-cdn-config/issues/14
It copies the code from `servicegovuk.vcl.erb` to redirect http://gov.uk/whatever to https://gov.uk/whatever and it sets a 10 second max-age to limit the damage if anything goes wrong.

I don't have a fastly account to test this against, so it's a little speculative.