Ensure the GOV.UK mirror stores the original version of the A/B test
Identify requests from the mirror's crawler worker, and set the request
header such that it only ever sees the original 'A' version of the
example A/B test.

This is in preparation for the real A/B test of education navigation. It
will affect many pages across the site, so it's important to make sure
that the mirror contains a self-consistent version of all the affected
pages.