Set a unique user ID as part of a Javascript detection experiment
We're re-running the experiment detailed in
https://gds.blog.gov.uk/2013/10/21/how-many-people-are-missing-out-on-javascript-enhancement/.

One of the changes we're making this time is to identify users rather
than requests. This commit sets the cookie value if it's absent and
writes it to a request header to preserve the value across requests
(after it's been stripped by Varnish, specifically).