Don't interrogate the fastly tls var when setting cookie
Instead we always set the cookie with the raw value in the
`tls.cookie.protocol` variable and this means we're future proofed against
new versions, and protected against the values changing.  We can use
analytics to create groups for secure and insecure values here if there
end up being lots of possible values.

We set the cookie as a session cookie (no max-age or expires) and don't
set it if it's already present.