Use query string to switch between navigation A/B test variants
The changes to education navigation will be tested by remote testers.
Unlike regular users, who should be assigned the A or B variant at
random, remote testers need to see a specific version of the A/B test.

Remote testers make their first visit to the site through a link that we
email to them. So a convenient way to place them in the A or B buckets
is to append a query string in that link that we send them, and identify
it when assigning buckets.

The query string will be `?ABTest-EducationNavigation=A`, i.e. one that
matches the A/B testing cookie. This isn't essential for this test, but
it will make it easier for us to commonise this config for future tests.