Fix setting of ABTest-EducationNavigation cookie for remote testers
Remote testers will be sent a URL like
https://www.gov.uk?ABTest-EducationNavigation=B to ensure that they see
a specific version of the A/B test. This was broken if they already had
an A/B testing cookie, because Fastly would not overwrite the existing
cookie.

This commit makes ensures the correct cookie is set if the URL contains
the A/B test query string, regardless of whether the user had an
existing cookie.