Add users to B group when going to a nav page
We are enabling the new navigation pages for everyone that goes directly
to them. The reason for this is to avoid people seeing 404s, and search
indexers index pages that might point people to 404s.

This is a complementary change where we assign users to the B group when
they go directly to a new navigation page, so they see a consistent
navigation throughout the site.