Allow query params on /education URLs
When adding smokey tests to verify we were assigning users to the
expected bucket, we realised we used a cache busting param to the
request. In doing so, the request wasn't recognised as being a request
to /education, meaning we weren't being assigned to the correct bucket.

This change makes sure we also allow for `?` after /education.