Add cookie for GOVUK-ABTest-Benchmarking
Background:

We are going to be running a new A/B test for the GOV.UK benchmarking
task. In this A/B test, which should last less than 1 week, we will be
adding a tracking library so we understand how users interact with the
new navigation pages we have exposed for Education content.

Changes:

In this commit we add rules to set the ABTest-Benchmarking cookie based
on a few things.

On `vcl_recv`:

- If the user visits GOV.UK with a parameter like
  `?ABTestBenchmarking=B`, we will use that value to set the header
  accordingly, so the application knows what variant to render (and set
  the `Vary` header appropriately);
- If the user already has a cookie set, we use that value in the
  header, so the application knows what variant to render (and set
  the `Vary` header appropriately);
- Finally, for anyone else, we always put them in the `A` bucket,
  because we only want to track users in the benchmarking test and no
  one else.

On `vcl_deliver`, on the first visit, we set the cookie and make sure
it expires after 7 days.