Add blue box A/B test for users in new navigation
This commit adds the header and cookie to all users that will be seeing
the existing EducationNavigation A/B test. We split those users into 2
groups: 50% will see things unchanged (control) and the other 50% will
see the new pattern, a blue box.

The cookie is set to expire after 30 days, which means that by then we
would have either removed the blue box or add it to everyone.