Add dynamic multivariate vcl templating
Adds the ability to define multivariate test configurations
in alphagov/fastly-configure by supplying a hash of nominators
for users (the denominator being the sum of nominators for the test).
eg:

- MyTest:
  - 3
  - 3
  - 3

The variants are assigned alphabetically so for this example
'A' variant would be assigned to one third of users,
likewise a third to 'B' and 'C'.