Refactor templating to use dictionary config
The drawback with storing bucket allocation config
in the templating layer is that the VCL needs to be
deployed when enabling/disabling and adjusting tests
by user allocation, so this commit moves the same logic
down into local VCL variables.
The downside of this is the allocation logic effectively
becomes untestable but the upside of dynamic configuration
is worth it.