Experiment with preloading the font
This extra header is used by supporting browsers to preload the font
(our biggest asset). This means that the browser can load the file in
parallel with the other work to load the page, it doesn't have to wait
before the DOM is loaded. This is an experiment to see if it makes a
difference in the performance of GOV.UK.

Initial tests on staging are inconclusive, and it's hard to test this
because staging doesn't have an SSL certificate so our speed test
software can't look at it. I intend to run this on GOV.UK for a few
weeks and use Speedcurve to monitor differences.

If we were going to use this, we'd obviously have to figure out a way
to make sure this URL doesn't go out of date. It hasn't changed in 4
years, but might soon.

https://trello.com/c/G1iddwUO/42-preload-assets-for-speed