Remove font preloading
This preloading of the font has been in place since Feb 1st
(https://github.com/alphagov/govuk-cdn-config/pull/64). This commit
rolls it back because:

- We are likely to be start implementing a different font loading
strategy, which makes use of WOFF and WOFF2 compression. At this point
we can't preload the font because this would likely result in extra
downloads (all users would get the WOFF2, even if they won't use it).
- Chrome seems to be loading the font twice. This could be caused by
the absence of the `crossorigin` in the preload header, but we don't
know for sure.