Test VCL against previously committed files
We intend to do some refactoring of the code in this repo. To make sure
that we don't unintentionally change the VCL that we send to Fastly,
this commit adds a test that generates the VCL for each environment and
service combination, and checks that against the VCL that we know is
good.

You can use `REGENERATE_EXPECTATIONS=1 bundle exec rspec` to regenerate
the files. That's useful if you've made changes that you know are fine.