Import code to deploy bouncer
This imports the code used to deploy the Fastly configuration for
Bouncer. The deploys are triggered (nightly) from Jenkins, which is
updated in https://github.com/alphagov/govuk-puppet/pull/8598.

The code is a straight copy-paste from alphagov/cdn-configs:

https://github.com/alphagov/cdn-configs/blob/6f45502ff2ef55c7c19870f0870
404f1c318f53f/fastly/configure_bouncer

The code at present doesn't conform to many of the standards we have on
GOV.UK, this will be fixed in follow up PRs. The goal of this PR is
just to make the CDN deploy system easier to understand by colocating
the code.

https://trello.com/c/y6MIgxjp