Import code to deploy the dictionaries
This imports the code to update the Fastly dictionaries. It currently
lives in cdn-configs (private repo), but most of it isn't secret. By
having the code here we'll have a single repo to configure the CDN.

The code in configure_dictionaries isn't up to any standard, but that
will be fixed in follow up PRs. This is similar to
https://github.com/alphagov/govuk-cdn-config/pull/92 and
https://github.com/alphagov/govuk-cdn-config/pull/89.

Last commit when importing the code:

https://github.com/alphagov/cdn-configs/tree/44464be3ec7a75308c96d824467
f6f2951689549

Part of addressing tech debt in CDN config:
https://trello.com/c/y6MIgxjp